backend/requirements.txt
python-telegram-bot[webhooks]==21.5
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
